static functions moved to ContextUtils
